Move `PositionAnimated`
