+++ packages/fuselage/src/components/AutoComplete/AutoComplete.js
@@ -4,12 +4,14 @@ import {
 } from '@rocket.chat/fuselage-hooks';
 import React, { useEffect, useRef, useMemo, useState } from 'react';
 
-import { Box, PositionAnimated, AnimatedVisibility } from '../Box';
+import AnimatedVisibility from '../AnimatedVisibility';
+import { Box } from '../Box';
 import Chip from '../Chip';
 import { Icon } from '../Icon';
 import { InputBox } from '../InputBox';
 import Margins from '../Margins';
 import { useCursor, Options } from '../Options';
+import PositionAnimated from '../PositionAnimated';
 
 const Addon = (props) => <Box rcx-autocomplete__addon {...props} />;
