--- packages/fuselage/src/components/Box/Position/position.stories.mdx.ignore
@@ -1,227 +1,0 @@
-import { useRef } from 'react';
-import { Meta, Canvas, ArgsTable, Story } from '@storybook/addon-docs/blocks';
-import { css } from '@rocket.chat/css-in-js';
-
-import { Position, Box, Field, Margins, ToggleSwitch, Tile } from '../..';
-
-<Meta title='Box/Modifiers/Position' parameters={{ jest: ['Box/spec'] }} />
-
-# Position
-
-<Canvas>
-  <Story name='top-start'>
-      {
-        () => {
-          const ref = useRef();
-          return <>
-            <Box display='flex' justifyContent='center' alignItems='center' height='400px' bg='neutral-800'>
-              <Box ref={ref} border='1px solid red' color='neutral-100'> anchor </Box>
-            </Box>
-            <Position anchor={ref} placement='top-start'>
-              <Tile width='150px'>hello</Tile>
-            </Position>
-          </>;
-        }
-      }
-  </Story>
-</Canvas>
-
-<Canvas>
-  <Story name='top-middle'>
-      {
-        () => {
-          const ref = useRef();
-          return <>
-            <Box display='flex' justifyContent='center' alignItems='center' height='400px' bg='neutral-800'>
-              <Box ref={ref} border='1px solid red' color='neutral-100'> anchor </Box>
-            </Box>
-            <Position anchor={ref} placement='top-middle'>
-              <Tile width='150px'>hello</Tile>
-            </Position>
-          </>;
-        }
-      }
-  </Story>
-</Canvas>
-
-<Canvas>
-  <Story name='top-end'>
-      {
-        () => {
-          const ref = useRef();
-          return <>
-            <Box display='flex' justifyContent='center' alignItems='center' height='400px' bg='neutral-800'>
-              <Box ref={ref} border='1px solid red' color='neutral-100'> anchor </Box>
-            </Box>
-            <Position anchor={ref} placement='top-end'>
-              <Tile width='150px'>hello</Tile>
-            </Position>
-          </>;
-        }
-      }
-  </Story>
-</Canvas>
-
-<Canvas>
-  <Story name='bottom-start'>
-      {
-        () => {
-          const ref = useRef();
-          return <>
-            <Box display='flex' justifyContent='center' alignItems='center' height='400px' bg='neutral-800'>
-              <Box ref={ref} border='1px solid red' color='neutral-100'> anchor </Box>
-            </Box>
-            <Position anchor={ref} placement='bottom-start'>
-              <Tile width='150px'>hello</Tile>
-            </Position>
-          </>;
-        }
-      }
-  </Story>
-</Canvas>
-
-<Canvas>
-  <Story name='bottom-middle'>
-      {
-        () => {
-          const ref = useRef();
-          return <>
-            <Box display='flex' justifyContent='center' alignItems='center' height='400px' bg='neutral-800'>
-              <Box ref={ref} border='1px solid red' color='neutral-100'> anchor </Box>
-            </Box>
-            <Position anchor={ref} placement='bottom-middle'>
-              <Tile width='150px'>hello</Tile>
-            </Position>
-          </>;
-        }
-      }
-  </Story>
-</Canvas>
-
-<Canvas>
-  <Story name='bottom-end'>
-    {
-      () => {
-        const ref = useRef();
-        return <>
-          <Box display='flex' justifyContent='center' alignItems='center' height='400px' bg='neutral-800'>
-            <Box ref={ref} border='1px solid red' color='neutral-100'> anchor </Box>
-          </Box>
-          <Position anchor={ref} placement='bottom-end'>
-            <Tile width='150px'>hello</Tile>
-          </Position>
-        </>;
-      }
-    }
-  </Story>
-</Canvas>
-
-<Canvas>
-  <Story name='right-start'>
-    {
-      () => {
-        const ref = useRef();
-        return <>
-          <Box display='flex' justifyContent='center' alignItems='center' height='400px' bg='neutral-800'>
-            <Box ref={ref} border='1px solid red' color='neutral-100'> anchor </Box>
-          </Box>
-          <Position anchor={ref} placement='right-start'>
-            <Tile width='150px'>hello</Tile>
-          </Position>
-        </>;
-      }
-    }
-  </Story>
-</Canvas>
-
-<Canvas>
-  <Story name='right-middle'>
-    {
-      () => {
-        const ref = useRef();
-        return <>
-          <Box display='flex' justifyContent='center' alignItems='center' height='400px' bg='neutral-800'>
-            <Box ref={ref} border='1px solid red' color='neutral-100'> anchor </Box>
-          </Box>
-          <Position anchor={ref} placement='right-middle'>
-            <Tile width='150px'>hello</Tile>
-          </Position>
-        </>;
-      }
-    }
-  </Story>
-</Canvas>
-
-<Canvas>
-  <Story name='right-end'>
-    {
-      () => {
-        const ref = useRef();
-        return <>
-          <Box display='flex' justifyContent='center' alignItems='center' height='400px' bg='neutral-800'>
-            <Box ref={ref} border='1px solid red' color='neutral-100'> anchor </Box>
-          </Box>
-          <Position anchor={ref} placement='right-end'>
-            <Tile width='150px'>hello</Tile>
-          </Position>
-        </>;
-      }
-    }
-  </Story>
-</Canvas>
-
-<Canvas>
-  <Story name='left-start'>
-    {
-      () => {
-        const ref = useRef();
-        return <>
-          <Box display='flex' justifyContent='center' alignItems='center' height='400px' bg='neutral-800'>
-            <Box ref={ref} border='1px solid red' color='neutral-100'> anchor </Box>
-          </Box>
-          <Position anchor={ref} placement='left-start'>
-            <Tile width='150px'>hello</Tile>
-          </Position>
-        </>;
-      }
-    }
-  </Story>
-</Canvas>
-
-<Canvas>
-  <Story name='left-middle'>
-    {
-      () => {
-        const ref = useRef();
-        return <>
-          <Box display='flex' justifyContent='center' alignItems='center' height='400px' bg='neutral-800'>
-            <Box ref={ref} border='1px solid red' color='neutral-100'> anchor </Box>
-          </Box>
-          <Position anchor={ref} placement='left-middle'>
-            <Tile width='150px'>hello</Tile>
-          </Position>
-        </>;
-      }
-    }
-  </Story>
-</Canvas>
-
-<Canvas>
-  <Story name='left-end'>
-    {
-      () => {
-        const ref = useRef();
-        return <>
-          <Box display='flex' justifyContent='center' alignItems='center' height='400px' bg='neutral-800'>
-            <Box ref={ref} border='1px solid red' color='neutral-100'> anchor </Box>
-          </Box>
-          <Position anchor={ref} placement='left-end'>
-            <Tile width='150px'>hello</Tile>
-          </Position>
-        </>;
-      }
-    }
-  </Story>
-</Canvas>
-
-<ArgsTable of={Position} />
