+++ packages/fuselage/src/components/Options/useCursor.ts
@@ -1,7 +1,7 @@
 import { useMutableCallback } from '@rocket.chat/fuselage-hooks';
 import { KeyboardEvent, useState } from 'react';
 
-import { AnimatedVisibility } from '../Box';
+import AnimatedVisibility from '../AnimatedVisibility';
 import type { OptionType } from './Options';
 import { useVisible } from './useVisible';
