+++ packages/fuselage/src/components/OptionsPaginated/OptionsPaginated.tsx
@@ -12,7 +12,8 @@ import React, {
 } from 'react';
 import { Virtuoso } from 'react-virtuoso';
 
-import { AnimatedVisibility, Box } from '../Box';
+import AnimatedVisibility from '../AnimatedVisibility';
+import { Box } from '../Box';
 import { CheckBox } from '../CheckBox';
 import Option from '../Options/Option';
 import Tile from '../Tile';
