+++ packages/fuselage/src/components/PositionAnimated/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PositionAnimated';
